+++ app/sheetBaker.ts
@@ -1,7 +1,5 @@
 
-const bakeRawSvg = async (svg, matchedIds, canvas) => {
-	// TODO: remove matched tokens, remove lilyond engraving signature
-
+const rasterizeSvg = async (svg, canvas) => {
 	const svgURL = "data:image/svg+xml," + encodeURIComponent(svg);
 	const image: any = await new Promise(resolve => {
 		const image = new Image();
@@ -19,6 +17,29 @@ const bakeRawSvg = async (svg, matchedIds, canvas) => {
 	//console.log("blob:", blob);
 
 	return URL.createObjectURL(blob);
+};
+
+
+const bakeRawSvg = (svg, matchedIds, canvas) => {
+	const dom = new DOMParser().parseFromString(svg, "text/xml");
+	const root: any = dom.childNodes[0];
+	//console.log("dom:", dom.childNodes[0]);
+
+	// remove lilypond engraving signature
+	for (const node of root.childNodes) {
+		if (node.tagName === "text") {
+			if (/www\.lilypond\.org/.test(node.textContent)) {
+				root.removeChild(node);
+				break;
+			}
+		}
+	}
+
+	// TODO: remove matched tokens
+
+	const doc = root.outerHTML;
+
+	return rasterizeSvg(doc, canvas);
 };